Removed Serial from constructor
Changed the code so that the constructor no longer needs to have the serial device. This is done because it's implemented as a static instance RS485 in RS485.cpp, where the port is hard-coded.

Instead there is now a constructor that takes no arguments, and a new function setserial() that allows the user to specify a hardwareSerial in their own code.